requirements.txt
Django==3.2.5
djangorestframework
python-environ
Pillow
six